// src/app/core/models/masters/organization/service-category.ts
export interface ServiceCategory {
    id: number;
    title: string;
    status: string;
    status_id :number;
    status_label: string;
}
export interface  ServiceCategoryPaginationResponse {
    current_page: number;
    total: number;
    per_page: number;
    last_page: number;
    from: number;
    data: ServiceCategory[];
}